Combined password controllers into one file
There were two password controllers for processing agent and chef. I merged all the functions into one page for convenience
