+++ SunDevilPizza/src/main/resources/com/pizza/sundevilpizza/passwordCheck.fxml
@@ -22,7 +22,7 @@
             <Font size="28.0" />
          </font>
       </PasswordField>
-      <Button fx:id="submit" layoutX="382.0" layoutY="334.0" mnemonicParsing="false" onAction="#submit" prefHeight="42.0" prefWidth="136.0" style="-fx-background-color: FA8072;" text="Submit" textFill="WHITE">
+      <Button fx:id="submit" layoutX="382.0" layoutY="334.0" mnemonicParsing="false" onAction="#submitProcess" prefHeight="42.0" prefWidth="136.0" style="-fx-background-color: FA8072;" text="Submit" textFill="WHITE">
          <font>
             <Font name="Comic Sans MS Bold" size="24.0" />
          </font>
